backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ezequiel Bitschin"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.4,<6.0.0",
    "djangorestframework>=3.16.0,<4.0.0",
    "returns>=0.26.0,<0.27.0",
    "django-filter>=25.1,<26.0",
    "pillow>=11.3.0,<12.0.0",
    "drf-spectacular>=0.28.0,<0.29.0",
    "dj-rest-auth>=7.0.1,<8.0.0",
    "django-allauth>=65.10.0,<66.0.0",
    "requests>=2.32.4,<3.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["blacar/"]

[tool.uv]
dev-dependencies = []